chore: Refactor mixin directory name
Add underscore as per feat/feature-loading
