+++ common/src/main/resources/assets/createcalamity/lang/en_us.json
@@ -1,4 +1,4 @@
 {
-  "death.attack.crushing": "%1$s was crushed to death",
-  "death.attack.crushing.player": "%1$s was crushed to death while fighting %1$s"
+  "death.attack.crushing": "%1$s got bonked",
+  "death.attack.crushing.player": "%1$s got bonked while fighting %1$s"
 }
